+++ newton/tests/test_examples.py
@@ -1191,6 +1191,14 @@ add_example_test(
     use_viewer=True,
 )
 
+add_example_test(
+    TestSoftbodyExamples,
+    name="softbody.example_softbody_limx_arap_bunnies_box",
+    devices=cuda_test_devices,
+    test_options={"num-frames": 80},
+    use_viewer=True,
+)
+
 
 class TestKaminoExamples(unittest.TestCase):
     pass